Refactor helpers package to apply
