+++ controllers/submariner/np_syncer_resources.go
@@ -23,7 +23,7 @@ import (
 
 	"github.com/go-logr/logr"
 	"github.com/submariner-io/submariner-operator/api/v1alpha1"
-	"github.com/submariner-io/submariner-operator/controllers/helpers"
+	"github.com/submariner-io/submariner-operator/controllers/apply"
 	"github.com/submariner-io/submariner-operator/pkg/discovery/network"
 	"github.com/submariner-io/submariner-operator/pkg/images"
 	"github.com/submariner-io/submariner-operator/pkg/names"
@@ -40,7 +40,7 @@ func (r *Reconciler) reconcileNetworkPluginSyncerDeployment(instance *v1alpha1.S
 ) error {
 	// Only OVNKubernetes needs networkplugin-syncer so far
 	if needsNetworkPluginSyncer(instance) {
-		_, err := helpers.ReconcileDeployment(instance, newNetworkPluginSyncerDeployment(instance,
+		_, err := apply.Deployment(instance, newNetworkPluginSyncerDeployment(instance,
 			clusterNetwork, names.NetworkPluginSyncerComponent), reqLogger, r.config.ScopedClient, r.config.Scheme)
 		return err
 	}
